pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "PyAsyncLogger"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiohttp",
    'numpy; python_version >= "3.10"',
    'pandas; python_version >= "3.10"',
]

[project.optional-dependencies]
gcp = ["google-cloud-logging"]

test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio",
    "coverage",
    "ruff",
    "sphinx",
    "tox",
]
dev = ["ruff", "pre-commit", "isort"]
lint = ["ruff"]
docs = ["sphinx"]
jupyter = ["jupyterlab"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 99
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
